requirements.txt
# Core LangChain and LangGraph dependencies
langchain==0.3.7
langchain-core==0.3.15
langchain-openai==0.2.5
langgraph==0.2.45

# Additional utilities
python-dotenv==1.0.0
